requirements.txt
Django==1.8
django-imagekit==3.2.7
django-meta==0.3.1
django-modeldict==1.4.1
MySQL-python==1.2.5
Pillow==2.9.0
pymongo==3.0.3
python-social-auth==0.2.13
djangorestframework==3.2.4
markdown==2.6.2
django-filter==0.11.0